MySql.Data.Entity 7.0.6-IR3 problems
MySql.Data.Entity 7.0.6-IR3 has this problem:
http://forums.mysql.com/read.php?38,652115,652115#msg-652115

As such, to support EF6 we have to stick with MySql.Data.Entity6.9.9.
It depends on MySql.Data6.9..9 which will not work with
MySql.Data7.0.6-IR31 which is what EFCore requires. So for now EFCore
support will have to be fixed to netstandard1.6. When the above issue is
fixed, we can support net451 also.