Support for backwards compatible endpoint names
In attempting to be compatible with the latest Spring Boot 2.0 changes
the endpoint names for trace, dump have changed to httptrace, threaddump
respectively. In order to support the new names while not breaking
backwards compatibility, this commit adds capability to have multiple Id
names by default. The configuration works as before, there will be an
additional request path that will be valid for trace and dump.

Addresses [#25]