Add Steeltoe.Common.Security package
Adds file based configuration provider to read PEM encoded certs and Keys and supports reload if the keys are rotated (i.e. Cloud Foundry)

Adds a IConfigureNamedOptions which reads from PEM config provider and converts the PEM files into PFX object and then into an X509Certificate2.

Adds a CertificateOptions that can be configured with the certificate and injected using the Options framework