Removed code that removes the "X-Forwarded-Proto" header.
For some reason in an ASP.Net WebForms application, the OWIN pipeline appears to restarts before a response is generated after the OpenIdConnectAuthenticationHandler.InvokeAsync. This results in ForwardedProtocolMiddleware being executed twice. Being as Steeltoe removes the "X-Forwarded-Proto" header, the scheme ends up being http instead of https resulting in a bad request_uri querystring parameter. It works fine in an ASP.NET MVC application. I can reproduce this and I am still working to isolate why the OWIN pipeline appears to restart. There is no harm in leaving "X-Forwarded-Proto" header and it resolves this issue.