Fix usages of IPAddress.Parse. (#223)
In ConsulServerUtils.FixIPv6Address, it is possible to get a non-valid IPAddress - ex. "localhost". This would cause an unnecessary exception, which was caught and discarded. Changed to use TryParse instead.

In InetUtils, the underlying method has an IPAddress, which it ToStrings, only to parse it again later. The change is to simply flow the IPAddress object.