RabbitMQ Test Fixes (#862)
* Fix TestDeferedAcks Unit Test

* Fix hang during RabbitMQ listener container shutdown

* Reduce number of sent messages TestDebatchByContainerPerformance

The TestDebatchByContainerPerformance() periodically fails when the RabbitMQ server is run in a docker container.  The underlying Socket connection would fail or drop causing the test to fail.  I suspect some sort of networking issue... but was not able to determine what exactly was happening... Reducing the number of messages sent seems to make this more stable.   Note: This test  does not fail when the RabbitMQ server is running on the same host as the tests themselves.

Co-authored-by: Tim Hess <thess@vmware.com>